split test into a separate job
