+++ .github/workflows/ci.yml
@@ -4,7 +4,7 @@ on: push
 
 jobs:
   check-format:
-    name: Check format & Test
+    name: Check format
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
@@ -13,6 +13,15 @@ jobs:
           node-version: 17
       - run: npm install
       - run: npm run check-format
+  unittests:
+    name: Check Tests
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions/setup-node@v3
+        with:
+          node-version: 17
+      - run: npm install
       - run: npm test
   build:
     name: Build
