+++ index.ts
@@ -1,5 +1,6 @@
 import { NextConfig } from 'next';
 import { NextWebVitalsMetric } from 'next/app';
+import sendMetrics from './sendMetrics';
 const _debounce = require('lodash/debounce');
 
 export function withAxiomProxy(nextConfig: NextConfig): NextConfig {
@@ -33,12 +34,10 @@ export function withAxiomProxy(nextConfig: NextConfig): NextConfig {
 }
 
 export function getCurrentPage() {
-  // TODO: find a way to mock this in tests
-  // return window.__NEXT_DATA__.page
-  return '/';
+  return window.__NEXT_DATA__.page
 }
 
-const debounceSendMetrics = _debounce(sendMetrics, 1000);
+const debounceSendMetrics = _debounce(() => sendMetrics(collectedMetrics), 1000);
 let collectedMetrics: NextWebVitalsMetric[] = [];
 
 // Usage:
@@ -47,19 +46,4 @@ export function reportWebVitals(metric: NextWebVitalsMetric) {
   metric['route'] = getCurrentPage();
   collectedMetrics.push(metric);
   debounceSendMetrics();
-}
-
-function sendMetrics() {
-  const url = '/axiom/web-vitals';
-  const body = JSON.stringify({
-    webVitals: collectedMetrics,
-  });
-
-  if (navigator.sendBeacon) {
-    navigator.sendBeacon(url, body);
-  } else {
-    fetch(url, { body, method: 'POST', keepalive: true });
-  }
-  // clear collectedMetrics
-  collectedMetrics = [];
 }
